requirements.txt
absl-py
astor
astropy
attrs
backcall
bleach
contextlib2
cycler
decorator
defusedxml
dm-sonnet
entrypoints
freeze
gast==0.3.3
grpcio
h5py==2.10.0
imageio
ipykernel
ipython
ipython-genutils
jedi
Jinja2
joblib
jsonschema
jupyter-client
jupyter-core
jupyterlab
jupyterlab-server
keras==2.3.1
Keras-Applications
Keras-Preprocessing
keras-tcn
kiwisolver
Markdown
MarkupSafe
matplotlib
mistune
mock
nbconvert
nbformat
notebook
numpy==1.18.1
pandocfilters
parso==0.7.0
patool
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
ptyprocess
Pygments
pyparsing
pyrsistent
python-dateutil
PyYAML
pyzmq
scikit-learn
scipy
semantic-version
Send2Trash
six
sklearn
tensorboard
tensorflow-estimator
tensorflow-gpu
tensorflow-probability
termcolor
terminado
testpath
tornado
traitlets
wcwidth
webencodings
Werkzeug
wrapt
